use existing "poor-man's virtualization" for lists
