+++ src/detail-panels/block-panel-generic/list-view.jsx
@@ -1,13 +1,9 @@
 // @ts-check
 /// <reference path="../../types.d.ts" />
 import { FormatTimestamp } from '../../common-components/format-timestamp';
-import { Visible } from '../../common-components/visible';
-import { useState } from 'react';
+import { ProgressiveRender } from '../../common-components/progressive-render';
 import { AccountShortEntry } from '../../common-components/account-short-entry';
 import { localise } from '../../localisation';
-
-const INITIAL_SIZE = 20;
-const GROW_BLOCK_SIZE = 29;
 
 /**
  * @param {{
@@ -15,36 +11,14 @@ const GROW_BLOCK_SIZE = 29;
  * }} _
  */
 export function ListView({ blocklist }) {
-  const [listSize, setListSize] = useState(INITIAL_SIZE);
-  const showSize = Math.min(blocklist.length, listSize);
-
   return (
     <ul className="block-list">
-      {blocklist.slice(0, showSize).map((block, index) => {
-        const entry = <ListViewEntry key={index} {...block} />;
-
-        return index < showSize - 1 ? (
-          entry
-        ) : (
-          <Visible
-            key={index}
-            rootMargin="0px 0px 300px 0px"
-            onVisible={handleBottomVisible}
-          >
-            {entry}
-          </Visible>
-        );
-      })}
+      <ProgressiveRender
+        items={blocklist}
+        renderItem={(item) => <ListViewEntry {...item} />}
+      />
     </ul>
   );
-
-  function handleBottomVisible() {
-    const incrementListSize = listSize + GROW_BLOCK_SIZE;
-    setListSize(incrementListSize);
-    if (incrementListSize > blocklist.length) {
-      // TODO: fetch more
-    }
-  }
 }
 
 /**
